Move tests from project_tests into project_search module
